Filters, Alignment Math + Tests
Created filters, which will be able to be stacked in the Display class
to perform operations on the displayed image to visualize stars and
luminance.
Created the Alignment class, containing the methods 'findPolarStar' and
the utility method 'rotate'. findPolarStar takes three points
(representing the same star on three images) and finds the point that
they appear to be rotating about, the North/South star for use in later
alignment of images.